fear(wallet,cockroach): add wallet db repos

+++ wallet/internal/adapters/db/sql/migrations/000002_account.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE accounts
 (
-  id      SERIAL PRIMARY KEY NOT NULL,
-  user_id INTEGER            NOT NULL,
-  name    varchar(20)        NOT NULL,
-  balance NUMERIC(12, 4)     NOT NULL DEFAULT 0.0000
+  id      BIGSERIAL PRIMARY KEY NOT NULL,
+  user_id BIGINT                NOT NULL,
+  name    varchar(20)           NOT NULL,
+  balance DOUBLE PRECISION      NOT NULL DEFAULT 0.0000
 );
 
 ALTER TABLE accounts
