+++ wallet/internal/adapters/db/sql/migrations/000004_transaction.up.sql
@@ -4,9 +4,9 @@ CREATE TABLE transactions
 (
   id                     uuid PRIMARY KEY          DEFAULT uuid_generate_v4(),
   type                   transaction_type NOT NULL,
-  amount                 NUMERIC(10, 2)   NOT NULL,
-  source_account_id      INTEGER REFERENCES accounts (id),
-  destination_account_id INTEGER REFERENCES accounts (id),
+  amount                 DOUBLE PRECISION NOT NULL,
+  source_account_id      BIGINT REFERENCES accounts (id),
+  destination_account_id BIGINT REFERENCES accounts (id),
   created_at             TIMESTAMP        NOT NULL DEFAULT NOW(),
   is_rolled_back         BOOLEAN          NOT NULL DEFAULT FALSE
 );
